kinship-ai-rewoo/requirements.txt
# LangGraph Dynamic Workflow Agent
# Python 3.10+ required

# ==================== CORE ====================
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# ==================== LANGGRAPH & LANGCHAIN ====================
langgraph>=1.0.5
langgraph-checkpoint-postgres>=3.0.2

langchain>=1.2.0
langchain-core>=1.2.5
langchain-openai>=1.1.6
langchain-google-genai>=4.1.2
langchain-mcp-adapters>=0.2.1
langchain-groq>=1.1.1

# ==================== LANGSMITH ====================
langsmith>=0.5.1

# ==================== DATABASE ====================
pymongo>=4.6.0
psycopg[binary]>=3.1.0

# ==================== HTTP & ASYNC ====================
aiohttp>=3.9.0
httpx>=0.25.0

# ==================== OPENAI ====================
openai>=1.0.0

# ==================== UTILITIES ====================
python-multipart>=0.0.6

# ==================== PRODUCTION ====================
gunicorn>=21.0.0

loguru
torch
silero_vad
groq
scipy